+++ back/src/domains/agency/use-cases/UpdateAgency.ts
@@ -1,6 +1,5 @@
 import {
   AgencyDto,
-  AgencyId,
   InclusionConnectedUser,
   agencySchema,
   errors,
@@ -9,6 +8,7 @@ import { TransactionalUseCase } from "../../core/UseCase";
 import { CreateNewEvent } from "../../core/events/ports/EventBus";
 import { UnitOfWork } from "../../core/unit-of-work/ports/UnitOfWork";
 import { UnitOfWorkPerformer } from "../../core/unit-of-work/ports/UnitOfWorkPerformer";
+import { throwIfNotAgencyAdminOrBackofficeAdmin } from "../../inclusion-connected-users/helpers/authorization.helper";
 
 export class UpdateAgency extends TransactionalUseCase<
   AgencyDto,
@@ -60,21 +60,3 @@ export class UpdateAgency extends TransactionalUseCase<
     ]);
   }
 }
-
-const throwIfNotAgencyAdminOrBackofficeAdmin = (
-  agencyId: AgencyId,
-  currentUser?: InclusionConnectedUser,
-): void => {
-  if (!currentUser) throw errors.user.unauthorized();
-  if (currentUser.isBackofficeAdmin) return;
-
-  const hasPermission = currentUser.agencyRights.some(
-    (agencyRight) =>
-      agencyRight.agency.id === agencyId &&
-      agencyRight.roles.includes("agency-admin"),
-  );
-
-  if (!hasPermission) {
-    throw errors.user.forbidden({ userId: currentUser.id });
-  }
-};
