+++ back/src/domains/core/feature-flags/use-cases/SetFeatureFlag.ts
@@ -3,7 +3,7 @@ import {
   SetFeatureFlagParam,
   setFeatureFlagSchema,
 } from "shared";
-import { throwIfNotAdmin } from "../../../inclusion-connected-users/helpers/throwIfIcUserNotBackofficeAdmin";
+import { throwIfNotAdmin } from "../../../inclusion-connected-users/helpers/authorization.helper";
 import { TransactionalUseCase } from "../../UseCase";
 import { UnitOfWork } from "../../unit-of-work/ports/UnitOfWork";
