+++ back/src/domains/inclusion-connected-users/helpers/authorization.helper.ts
@@ -1,4 +1,9 @@
-import { InclusionConnectDomainJwtPayload, errors } from "shared";
+import {
+  AgencyId,
+  InclusionConnectDomainJwtPayload,
+  InclusionConnectedUser,
+  errors,
+} from "shared";
 import { UserOnRepository } from "../../core/authentication/inclusion-connect/port/UserRepository";
 import { UnitOfWork } from "../../core/unit-of-work/ports/UnitOfWork";
 import { getIcUserByUserId } from "./inclusionConnectedUser.helper";
@@ -15,4 +20,22 @@ export const throwIfIcUserNotBackofficeAdmin = async (
 export const throwIfNotAdmin = (user: UserOnRepository | undefined) => {
   if (!user) throw errors.user.unauthorized();
   if (!user.isBackofficeAdmin) throw errors.user.forbidden({ userId: user.id });
+};
+
+export const throwIfNotAgencyAdminOrBackofficeAdmin = (
+  agencyId: AgencyId,
+  currentUser?: InclusionConnectedUser,
+): void => {
+  if (!currentUser) throw errors.user.unauthorized();
+  if (currentUser.isBackofficeAdmin) return;
+
+  const hasPermission = currentUser.agencyRights.some(
+    (agencyRight) =>
+      agencyRight.agency.id === agencyId &&
+      agencyRight.roles.includes("agency-admin"),
+  );
+
+  if (!hasPermission) {
+    throw errors.user.forbidden({ userId: currentUser.id });
+  }
 };
